feat: add traefik func
