+++ scripts/autocompletes/vfd.zsh
@@ -36,6 +36,7 @@ _vfd() {
         case $line[1] in
             (up)
 _arguments "${_arguments_options[@]}" \
+'--no-traefik[Skips the traefik domain routing]' \
 '--no-detach[Runs the docker compose command without detaching]' \
 '-h[Print help]' \
 '--help[Print help]' \
@@ -43,6 +44,7 @@ _arguments "${_arguments_options[@]}" \
 ;;
 (down)
 _arguments "${_arguments_options[@]}" \
+'--no-traefik[Skips the traefik domain routing]' \
 '-h[Print help]' \
 '--help[Print help]' \
 && ret=0
